Split openers into their own file; rename
